Disabled microphone permission check in FCM `onCallInvite` (see details)
Foreground services requiring privacy permission e.g. microphone or camera are required to be started in the foreground. Since we're using the Telecom's PhoneAccount, we don't directly require microphone access.

Further, microphone access is always denied if the app requiring microphone access via a Foreground service is in the background (by design).