Updated on call check with static public field (see details)
Due to `TelecomManager.isInCall` checks for any active `ConnectionService` (Connections), the function returns true even if there is an active cellular call.

A first alternative is providing the active call handle via a `callSid` field, however this is unreliable if app is started when a call is ongoing. Init functions can be added, however this is cumbersome.

A hack'ish solution seems to be preferred by making the `TVConnectionService.activeConnections` and helper functions static to provide easy of access in the Plugin.